refactor: refactor Navbar components, move SearchModal to Navbar folder
