+++ frontend/src/components/server/navbar/ProfileDropdown.tsx
@@ -1,63 +1,80 @@
 import React, { FunctionComponent } from 'react'
+import { useAppDispatch, useAppSelector } from 'redux/hooks'
+
+import { selectCurrentUser } from 'components/auth/redux/auth.slice'
+import { signout } from 'components/auth/redux/auth.thunk'
+
 import { Dropdown } from 'semantic-ui-react'
-import { userOptions } from 'utils/appConst'
 
-interface ProfileDropdownProps {
-    username: string
-    avatarUrl: string
-    userStatus: string
-    handleSignout: Function
-}
+interface ProfileDropdownProps {}
 
-const ProfileDropdown: FunctionComponent<ProfileDropdownProps> = ({
-    username,
-    avatarUrl,
-    userStatus,
-    handleSignout,
-}) => {
-    const menuOptions = userOptions
+const ProfileDropdown: FunctionComponent<ProfileDropdownProps> = () => {
+    const dispatch = useAppDispatch()
+    const currentUser = useAppSelector(selectCurrentUser)
 
-    const onSignoutClick = () => {
-        handleSignout()
+    const menuOptions = {
+        userProfile: {
+            text: 'Profile',
+            disabled: true,
+        },
+        userSetting: {
+            text: 'Preferences',
+        },
+        signout: {
+            text: 'Sign out',
+        },
     }
 
-    return (
-        <div className="ml-auto mx-4 bg-white rounded-md max-h-6 w-6">
-            <Dropdown
-                trigger={
-                    <span className="flex">
-                        <img src={avatarUrl} alt="avt" className="rounded-md" />
-                    </span>
-                }
-                icon={null}
-                id="user-settings"
-            >
-                <Dropdown.Menu className="left text-black">
-                    <div className="flex items-center justify-start m-4 pr-32">
+    const onSignoutClick = async () => {
+        await dispatch(signout())
+    }
+
+    if (currentUser) {
+        const { photoUrl, username, status } = currentUser
+        return (
+            <div className="ml-auto mx-4 bg-white rounded-md max-h-6 w-6">
+                <Dropdown
+                    trigger={
                         <span className="flex">
                             <img
-                                className="max-h-12 w-12 rounded-md"
-                                src={avatarUrl}
+                                src={photoUrl}
                                 alt="avt"
+                                className="rounded-md"
                             />
                         </span>
-                        <div className="pl-4">
-                            <h3 className="font-semibold">{username}</h3>
+                    }
+                    icon={null}
+                    id="user-settings"
+                >
+                    <Dropdown.Menu className="left text-black">
+                        <div className="flex items-center justify-start m-4 pr-32">
+                            <span className="flex">
+                                <img
+                                    className="max-h-12 w-12 rounded-md"
+                                    src={photoUrl}
+                                    alt="avt"
+                                />
+                            </span>
+                            <div className="pl-4">
+                                <h3 className="font-semibold">{username}</h3>
+                            </div>
+                            <h4>{status}</h4>
                         </div>
-                        <h4>{userStatus}</h4>
-                    </div>
-                    <Dropdown.Divider />
-                    <Dropdown.Item {...menuOptions[0]} />
-                    <Dropdown.Item {...menuOptions[1]} />
-                    <Dropdown.Divider />
-                    <Dropdown.Item
-                        {...menuOptions[2]}
-                        onClick={onSignoutClick}
-                    />
-                </Dropdown.Menu>
-            </Dropdown>
-        </div>
-    )
+                        <Dropdown.Divider />
+                        <Dropdown.Item {...menuOptions.userProfile} />
+                        <Dropdown.Item {...menuOptions.userSetting} />
+                        <Dropdown.Divider />
+                        <Dropdown.Item
+                            {...menuOptions.signout}
+                            onClick={onSignoutClick}
+                        />
+                    </Dropdown.Menu>
+                </Dropdown>
+            </div>
+        )
+    } else {
+        return null
+    }
 }
 
 export default ProfileDropdown
